+++ Main.java
@@ -1,50 +1,24 @@
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
 import java.util.Scanner;
-import java.util.LinkedList;
 
 public class Main {
-    public static void main(String args[]) {
+    public static void main() {
         Scanner sc = new Scanner(System.in);
-        LinkedList<Integer> deque = new LinkedList<Integer>();
-
-        int count = 0;
-        int N = sc.nextInt();
-        int M = sc.nextInt();
-        
-        for (int i = 0; i < N; i++) {
-            deque.add(i + 1);
-        }
-        int[] arr = new int[N];
-        for (int i = 0; i < N; i++) {
-            arr[i] = sc.nextInt();
-        }
-        
-        for (int i = 0; i < M; i++) {
-            int target_idx = deque.indexOf(arr[i]);
-            int half_idx;
-
-            if (deque.size() % 2 == 0) {
-                half_idx = deque.size() / 2 - 1;
+        List<Integer> list = new ArrayList<>();
+        Random rand = new Random();
+        int input = sc.nextInt();
+        int num;
+        for (int i = 0; i < input; i++) {
+            num = rand.nextInt(i);
+            if (num <= i) {
+                list.add(0, num);
             }
             else {
-                half_idx = deque.size() / 2;
+                list.add(i, num);
             }
-
-            if (target_idx <= half_idx) {
-                for (int j = 0; j < target_idx; j++) {
-                    int temp = deque.pollFirst();
-                    deque.offerLast(temp);
-                    count++;
-                }
-            }
-            else {
-                for (int j = 0; j < deque.size() - target_idx; j++) {
-                    int temp = deque.pollLast();
-                    deque.offerFirst(temp);
-                    count++;
-                }
-            }
-            deque.pollFirst();
         }
-        System.out.println(count);
+        System.out.println(list);
     }
 }
